Find memory leak
Use short for year values instead of casted int.

+++ mobitopp/src/main/java/edu/kit/ifv/mobitopp/util/panel/HouseholdInfo.java
@@ -3,7 +3,7 @@ package edu.kit.ifv.mobitopp.util.panel;
 class HouseholdInfo {
 
 	public long household_number;
-	public int year;
+	public short year;
 	public int area_type;
 	public int household_size;
 	public int domcode;
