+++ mobitopp/src/main/java/edu/kit/ifv/mobitopp/util/panel/PaneldataReader.java
@@ -117,7 +117,7 @@ public class PaneldataReader {
 		String[] field = line.split(";");
 
 		info.household.household_number 	= java.lang.Long.parseLong(field[columnNames.get("id")]);
-		info.household.year 							= java.lang.Integer.parseInt(field[columnNames.get("year")]);
+		info.household.year 							= java.lang.Short.parseShort(field[columnNames.get("year")]);
 		info.household.area_type 					= java.lang.Integer.parseInt(field[columnNames.get("areatype")]);
 		info.household.household_size 		= java.lang.Integer.parseInt(field[columnNames.get("size")]);
 		info.household.domcode 						= java.lang.Integer.parseInt(field[columnNames.get("hhtype")]);
